feat(demo): 添加系统信息查询功能并优化数据统计模型
- 新增 SysInfoController 和 SysInfoService 类，实现系统信息查询功能
- 修改 SysInfoStat 类，将 count 字段类型从 Integer改为 Long，以支持更大的统计数值

+++ src/main/java/cn/zhaofd/demojpaweb/demo/dto/SysInfoStat.java
@@ -5,9 +5,9 @@ package cn.zhaofd.demojpaweb.demo.dto;
  */
 public class SysInfoStat {
     private String name;
-    private Integer count;
+    private Long count;
 
-    public SysInfoStat(String name, Integer count) {
+    public SysInfoStat(String name, Long count) {
         this.name = name;
         this.count = count;
     }
@@ -20,11 +20,11 @@ public class SysInfoStat {
         this.name = name;
     }
 
-    public Integer getCount() {
+    public Long getCount() {
         return count;
     }
 
-    public void setCount(Integer count) {
+    public void setCount(Long count) {
         this.count = count;
     }
 }
